Move view_running.cpp to view_text.cpp
